backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
Pillow==10.3.0
exifread==3.0.0
watchdog==4.0.0
aiosqlite==0.20.0
python-multipart==0.0.9
